Add fastapi route, move server to main, update page
